+++ global_segment_map_node/CMakeLists.txt
@@ -42,7 +42,6 @@ endif()
 
 cs_add_library(${PROJECT_NAME}_library
   src/controller.cpp
-  src/feature_ros_tools.cpp
 )
 target_link_libraries(${PROJECT_NAME}_library ${catkin_LIBRARIES} ${approxmvbb_catkin_LIBRARIES})
